moved pre-v1 constraints/types to legacy module
