+++ tableschema/legacy/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
+from . import constraints
+from . import types
