+++ tableschema/legacy/constraints/__init__.py
