+++ tableschema/legacy/constraints/enum.py
@@ -4,28 +4,28 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-from .. import exceptions
+from ... import exceptions
 
 
 # Module API
 
-def check_minLength(name, value, min_length):
-    """Length constraint.
+def check_enum(name, value, enum):
+    """Enum constraint.
 
-    Supported types: sting, array, object.
+    Check value is one from enumerable type.
 
     Args:
         name (str): field name
-        value (str/list/dict): field value
-        min_length (str/list/dict): min length to check
+        value (any): field value
+        enum (any[]): enum to check against
 
     Raises:
         TypeError: for non supported type
         ConstraintError: if check is failed
 
     """
-    if len(value) < min_length:
+    if value not in enum:
         raise exceptions.ConstraintError(
-            "The field '{0}' must have a minimum length of {1}"
-            .format(name, min_length))
+            "The value for field '{0}' must be in the enum array"
+            .format(name))
     return True
