+++ tableschema/legacy/constraints/maximum.py
@@ -4,7 +4,7 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-from .. import exceptions
+from ... import exceptions
 
 
 # Module API
