+++ tableschema/legacy/types/any.py
@@ -4,7 +4,7 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-from .. import helpers
+from ... import helpers
 from . import base
