+++ tableschema/legacy/types/base.py
@@ -7,9 +7,9 @@ from __future__ import unicode_literals
 from six import add_metaclass
 from functools import partial
 from abc import ABCMeta, abstractmethod
-from .. import compat
-from .. import helpers
-from .. import exceptions
+from ... import compat
+from ... import helpers
+from ... import exceptions
 from .. import constraints
