+++ tableschema/legacy/types/boolean.py
@@ -4,8 +4,8 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-from .. import exceptions
-from .. import helpers
+from ... import exceptions
+from ... import helpers
 from . import base
