+++ tableschema/legacy/types/date.py
@@ -4,22 +4,21 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-import time
 import datetime
 from future.utils import raise_with_traceback
 from dateutil.parser import parse as date_parse
-from .. import exceptions
-from .. import helpers
+from ... import exceptions
+from ... import helpers
 from . import base
 
 
 # Module API
 
-class TimeType(base.JTSType):
+class DateType(base.JTSType):
 
     # Public
 
-    name = 'time'
+    name = 'date'
     null_values = helpers.NULL_VALUES
     supported_constraints = [
         'required',
@@ -29,10 +28,11 @@ class TimeType(base.JTSType):
         'maximum',
     ]
     # ---
-    python_type = datetime.time
-    ISO8601 = '%H:%M:%S'
-    raw_formats = ['HH/MM/SS']
-    py_formats = ['%H:%M:%S']
+    python_type = datetime.date
+    formats = ('default', 'any', 'fmt')
+    ISO8601 = '%Y-%m-%d'
+    raw_formats = ['DD/MM/YYYY', 'DD/MM/YY', 'YYYY/MM/DD']
+    py_formats = ['%d/%m/%Y', '%d/%m/%y', '%Y/%m/%d']
     format_map = dict(zip(raw_formats, py_formats))
 
     def cast_default(self, value, fmt=None):
@@ -41,11 +41,9 @@ class TimeType(base.JTSType):
             return value
 
         try:
-            struct_time = time.strptime(value, self.ISO8601)
-            return datetime.time(
-                struct_time.tm_hour, struct_time.tm_min, struct_time.tm_sec)
+            return datetime.datetime.strptime(value, self.ISO8601).date()
         except (TypeError, ValueError) as e:
-            raise_with_traceback(exceptions.InvalidTimeType(e))
+            raise_with_traceback(exceptions.InvalidDateType(e))
 
     def cast_any(self, value, fmt=None):
 
@@ -53,9 +51,9 @@ class TimeType(base.JTSType):
             return value
 
         try:
-            return date_parse(value).time()
+            return date_parse(value).date()
         except (TypeError, ValueError) as e:
-            raise_with_traceback(exceptions.InvalidTimeType(e))
+            raise_with_traceback(exceptions.InvalidDateType(e))
 
     def cast_fmt(self, value, fmt=None):
 
@@ -63,6 +61,6 @@ class TimeType(base.JTSType):
             return value
 
         try:
-            return datetime.datetime.strptime(value, fmt).time()
+            return datetime.datetime.strptime(value, fmt).date()
         except (TypeError, ValueError) as e:
-            raise_with_traceback(exceptions.InvalidTimeType(e))
+            raise_with_traceback(exceptions.InvalidDateType(e))
