+++ tableschema/legacy/types/datetime.py
@@ -7,8 +7,8 @@ from __future__ import unicode_literals
 import datetime
 from future.utils import raise_with_traceback
 from dateutil.parser import parse as date_parse
-from .. import exceptions
-from .. import helpers
+from ... import exceptions
+from ... import helpers
 from . import base
