+++ tableschema/legacy/types/duration.py
@@ -8,8 +8,8 @@ import isodate
 from future.utils import raise_with_traceback
 
 from . import base
-from .. import exceptions
-from .. import helpers
+from ... import exceptions
+from ... import helpers
 
 
 # Module API
