+++ tableschema/legacy/types/geopoint.py
@@ -7,9 +7,9 @@ from __future__ import unicode_literals
 import json
 import decimal
 from future.utils import raise_with_traceback
-from .. import exceptions
-from .. import helpers
-from .. import compat
+from ... import exceptions
+from ... import helpers
+from ... import compat
 from . import base
