+++ tableschema/legacy/types/number.py
@@ -9,9 +9,9 @@ import six
 import decimal
 import unicodedata
 from future.utils import raise_with_traceback
-from .. import exceptions
-from .. import helpers
-from .. import compat
+from ... import exceptions
+from ... import helpers
+from ... import compat
 from . import base
