+++ tableschema/legacy/types/object.py
@@ -6,8 +6,8 @@ from __future__ import unicode_literals
 
 import json
 from future.utils import raise_with_traceback
-from .. import exceptions
-from .. import helpers
+from ... import exceptions
+from ... import helpers
 from . import base
