+++ tests/legacy/test_types.py
@@ -8,8 +8,9 @@ import isodate
 import pytest
 from datetime import datetime, date, time, timedelta
 from decimal import Decimal
-from tableschema import types, exceptions
-from . import base
+from tableschema import exceptions
+from tableschema.legacy import types
+from .. import base
 
 
 class TestString(base.BaseTestCase):
